fix: Only list owned accounts when adding to space
