+++ apps/web/src/features/spaces/hooks/useSpaceSafes.tsx
@@ -1,37 +1,13 @@
 import { useOrganizationSafesGetV1Query } from '@safe-global/store/gateway/AUTO_GENERATED/organizations'
 import { useCurrentSpaceId } from 'src/features/spaces/hooks/useCurrentSpaceId'
-import type { AllSafeItems } from '@/features/myAccounts/hooks/useAllSafesGrouped'
+import { _buildSafeItems, type AllSafeItems } from '@/features/myAccounts/hooks/useAllSafesGrouped'
 import { useAllSafesGrouped } from '@/features/myAccounts/hooks/useAllSafesGrouped'
 import { useAppSelector } from '@/store'
 import { selectOrderByPreference } from '@/store/orderByPreferenceSlice'
 import { getComparator } from '@/features/myAccounts/utils/utils'
 import { useMemo } from 'react'
-import type { SafeItem } from '@/features/myAccounts/hooks/useAllSafes'
-import { selectAllAddressBooks, type AddressBookState } from '@/store/addressBookSlice'
+import { selectAllAddressBooks } from '@/store/addressBookSlice'
 import { isAuthenticated } from '@/store/authSlice'
-
-function _buildSafeItems(safes: Record<string, string[]>, allSafeNames: AddressBookState): SafeItem[] {
-  const result: SafeItem[] = []
-
-  for (const chainId in safes) {
-    const addresses = safes[chainId]
-
-    addresses.forEach((address) => {
-      const name = allSafeNames[chainId]?.[address]
-
-      result.push({
-        chainId,
-        address,
-        isReadOnly: false,
-        isPinned: false,
-        lastVisited: 0,
-        name,
-      })
-    })
-  }
-
-  return result
-}
 
 export const useSpaceSafes = () => {
   const spaceId = useCurrentSpaceId()
